be/requirements.txt
django==3.2
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
djoser==2.1.0
django_cors_headers==3.7.0